--- app/src/main/java/com/mit/offroader/ui/fragment/map/MapViewModel.kt
@@ -1,11 +1,0 @@
-package com.mit.offroader.ui.fragment.map
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class MapViewModel : ViewModel() {
-    private var _mapUiState = MutableLiveData<MapUiState>()
-
-    val mapUiState : LiveData<MapUiState> = _mapUiState
-}
