+++ app/src/main/java/com/mit/offroader/ui/fragment/map/SanMapUiState.kt
@@ -1,5 +1,5 @@
 package com.mit.offroader.ui.fragment.map
 
-data class MapUiState(
+data class SanMapUiState(
     val value: String, // 임시로 넣어둔 상수입니당 :)
 )
