+++ app/src/main/java/com/mit/offroader/ui/fragment/map/SanMapViewModel.kt
@@ -1,0 +1,11 @@
+package com.mit.offroader.ui.fragment.map
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SanMapViewModel : ViewModel() {
+    private var _San_mapUiState = MutableLiveData<SanMapUiState>()
+
+    val sanMapUiState : LiveData<SanMapUiState> = _San_mapUiState
+}
