+++ app/src/main/res/drawable/map_search_edge.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="#FFFFFF" />
+            <stroke android:width="1dp" android:color="@color/black"/>
+            <corners android:radius="25dp" />
+        </shape>
+    </item>
+
+</selector>
